add print last container logs if deploy failed
